pyproject.toml
[project]
name = "chat_pdf"
version = "0.1.0"
description = "Chat PDF"
authors = [
    {name = "amandaozava", email = "amanda.ozava@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "streamlit (>=1.45.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "langchain-community (>=0.3.23,<0.4.0)",
    "openai (>=1.78.0,<2.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "streamlit-chat (>=0.1.1,<0.2.0)",
    "langchain-openai (>=0.3.16,<0.4.0)",
    "chromadb (>=1.0.12,<2.0.0)",
    "protobuf (==3.20.3)",
    "unstructured[all] (>=0.17.2,<0.18.0)",
    "docx (>=0.2.4,<0.3.0)",
    "python-docx (>=1.1.2,<2.0.0)",
    "docx2txt (>=0.9,<0.10)",
    "lint (>=1.2.1,<2.0.0)"
]
[tool.poetry.group.dev.dependencies]
ruff = "^0.12.1"
[tool.ruff]
line-length = 100
target-version = "py310"  # Altere se estiver usando outro Python
exclude = ["venv", ".venv", "migrations", "__pycache__"]

